cc1: do not attempt copy propagation in OPT_MCH_MBZ
this did not achieve much (eliminated 3 insns out of ~70k in
regression) and made a bit of a mess of the rules for register
substitution. besides, once the graph allocator is reworked to
permit mixed-sized register use we'll be able to interleave
this pass with allocation, and the latter will do coalescing.

re-instated the rule for insn_substitute_reg that the src must
not be a machine reg, since OPT_MCH_MBZ was the only client
which tried this. simplified the implementation accordingly.

Signed-off-by: Charles E. Youse <charles@gnuless.org>