cc1/graph.c: do spills using the node symbol's type
previously we simply assumed the worst-case for the node's
cast class for the spill register's type, then attempted to
cater loads and stores based on the operand types of the
insns that referred to the spilled node. (the intent was to
maximize fusing possibilities.) now, instead, we make use of
the fact that the node's associated symbol has a type that
is large enough for every use of the reg, and blindly load
and store using that size. this is simpler, not fraught with
loads of subtle bugs, makes it easier to ease the size rules
on the MCH IR, and has negligible effect on fusing.

Signed-off-by: Charles E. Youse <charles@gnuless.org>