cc1/graph.c: record blocks where nodes appear
when computing spill costs, record which blocks nodes
actually appear in. the immediate benefit is that this
reduces the work to rewrite a node once we decide to
spill it, since we need not scan blocks where the node
doesn't appear, but the main reason we add this is to
provide a datum for some heuristics to be added shortly.

Signed-off-by: Charles E. Youse <charles@gnuless.org>