cc1/lower.c: rename lower_more() -> desynth() and reorder
this function does nothing but rewrite awkward fake MCH
insns for output purposes, and so it should be scheduled
immediately before output, not before the OPT_MCH_FINAL
passes. so reorder, make a minor change to OPT_MCH_MBZ to
account for reordering, and give it a more sensible name.

Signed-off-by: Charles E. Youse <charles@gnuless.org>