cc1/opt.c: expand tail merging to machine IR
tail merging is a simple and effective way
to reduce code size, and there's no reason
to limit it to LIR. we rename OPT_LIR_MERGE
to OPT_MERGE and reclassify to OPT_ANY_PASSES.
this removes ~1.2% of insns from regression.

bug fixed: merge0() was appending migrated insns
to the merge block, which reverses the order if
multiple insns are merged. this bug did not rear
its head when we were only merging the LIR, but
only by luck. changed to prepend insns instead.

Signed-off-by: Charles E. Youse <charles@gnuless.org>