yacc: final ANSI modifications (for now)
remove remaining `register' declarations (which are
useless on modern implementations, especially when,
as in this case, *everything* is declared register).

add a prototype for allocate() (defined in main.c).

these are the final modernizations for the moment,
though they are certainly places where better use
of ANSI C code be made; but at least now we have
prototypes/new-style definitions for everything.

Signed-off-by: Charles E. Youse <charles@gnuless.org>