ar: port of V7 `ar' utility
the archive format has changed, though only in minor
details (size of fields). comments have been added to
the code and it's been modernized for ANSI and POSIX.

i've refrained from making too many structural changes,
though it should be revisited at a future date. the v7
`ar' was itself a conversion from the earlier v6 version
which was written in assembly, so the c is a bit rough.

Signed-off-by: Charles E. Youse <charles@gnuless.org>