cc1/graph.c: range_interf(): do not check for Y->X
there is no need to check for copies from Y to X when
deciding whether Y interferes with X; since that DEFs
X, that will not be in the range `r' we are examining,
as a DEF would (by definition) begin a new live range.

no regressions with the test removed.

cleaned up the comments again to better reflect
the definition of interference as we employ it.

Signed-off-by: Charles E. Youse <charles@gnuless.org>