cc1/graph.c: coalesce non-destructive sign extensions
under some circumstances, a register can hold two
logical values at once: a value and its extension.

e.g., in a sequence like

                  movslq %i100l, %i101q
                  movb (%%i101q), %i102b

as long as %i100 is backed by a symbol that is no larger
than an int (meaning its high bits are guaranteed junk),
and %i100 and %i101 do not otherwise interfere [per our
usual graph-coloring rules], then we can coalesce them.

this reduces register pressure, which generally improves
the code, but there are a few negative regressions which
are caused by poor interactions with the spill logic. as
the coalesced node is larger than the original, if it is
spilled it is spilled as the larger size, which inhibits
fusing; this can be addressed elsewhere at a later date.

Signed-off-by: Charles E. Youse <charles@gnuless.org>