as: add templates for AND insn
one line is commented out of the regression test:

`as' quite rightly notes that the immediate is out of range.
`gas' has let it slide, but clearly the compiler should not
be emitting this immediate as-is. this must be investigated
but is not related to the assembler or template for AND.

also, added regress/a.out to the .gitignore. just an oversight.

Signed-off-by: Charles E. Youse <charles@gnuless.org>