cc1/gen.c: normalize mask during field insertion
the output for gen.s (purely coincidental that the bug
arose from code ALSO in gen.c) included:

		andl $-4026531841,%ecx

which `as' (but not `gas') complained about, because
the constant is not a valid 32-bit immediate value.
this was caused by inverting 0xF0000000 as a long,
resulting in 0xFFFFFFFF0FFFFFFF (in decimal above).
we must normalize this value to cut the high bits.

`gas' simply truncated the value, but `as' does not.

there is a bunch of regression noise, since the high
bits are now clear where they were not before, though
(aside from fixing the above snafu) no real changes.

Signed-off-by: Charles E. Youse <charles@gnuless.org>