cc1/zlq.c: compute new state into a temporary
we were computing the new state directly into the state
variable (b->zlq), so we fell into the usual trap when
a block was its own predecessor (nuking state before we
were done with it) causing us to miss some opportunities.

no regressions, but other code (sieve.c) shows improvement.
(should we add some benchmarks to the regression suite?)

this will become more important when we rewrite SLQs as ZLQs.
also, the fact that we're not doing proper data-flow analysis
in this pass will also come back to haunt us; we need to use a
real lattice to carry data around loops. revisit in the future.

Signed-off-by: Charles E. Youse <charles@gnuless.org>