cross.sh: expand cross-build script
we now build two versions of the cross tools: one which
builds binaries for linux using gnu binutils, and another
which builds binaries for tahoe using its own binary tools.
the former is used primarily to build the latter, and the
latter is used to cross-build the kernel and filesystem.

also took the time to reorganize the target filesystem tree.
there's no reason to have a /usr tree any more, since separate
root and /usr filesystems is a thing of the past. so we'll put
binaries in /bin, libraries and support binaries in /lib, the
includes straight in /include, etc. updated the cross.sh, cc.c
and the compiler's first pass regression suite to reflect this.

Signed-off-by: Charles E. Youse <charles@gnuless.org>