big changes to cross.sh bootstrapping procedure
the lex makefile has been rewritten to conform to our
standard, and it is now built with tahoe cc. relatively
minor changes have been made to `as' so that it is now
buildable with tahoe lex and yacc, and cross.sh now does.

because these are older versions of lex and yacc,
this makes it difficult to build `as' as a native
tool on Linux with its native toolchain; tahoe's
first-stage tools built in cross.sh MUST be used.
(this is no different from ar, ld, nm which also
can't be built with gcc on linux because of headers)

Signed-off-by: Charles E. Youse <charles@gnuless.org>