boot: eliminate idle_stack, shuffle addresses
we are always going to enter the kernel on the boot
stack, so remove the parameterized idle_stack. move
PER_CPU to where the idle stack used to be, to give
both the per_cpu[] array and the E820 map more room.

shuffle the addresses of zero-page temporaries down
to give the boot stack a little more room, too.

Signed-off-by: Charles E. Youse <charles@gnuless.org>