cc1/insn.c: no (%rip) with absolute addresses
previously, when given an expression like

	*((char *) 0x1000) = 65;

the compiler would emit something like

	movb $65, 0x1000(%rip)

but this is not correct; `as' assumes that
the user really means an offset of 0x1000
from %rip. what should be emitted is

	movb $65, 0x1000

so now we check to see if there's a symbol
before attaching the (%rip) suffix.

Signed-off-by: Charles E. Youse <charles@gnuless.org>