cc1/tree.c: add shiftmask0() tree optimization
the ATOM will automatically apply a mask to the bitcount
of a shift operation, so if the user explicitly applies
that mask, their mask is redundant and can be eliminated.

we change the semantics of E_SHR/E_SHL/E_SHRASG/E_SHLASG
trees to imply this mask, and rewrite trees to eliminate
and duplicate mask supplied by the user. minor updates to
constant-folding of trees and in the LIR to honor the new
semantics. also mask constants emitted w/ MCH shift insns.
this last change is only tangentially related, but it fixes
a problem where the assembler would choke if given, e.g.,
`i << 1024' since `shll $1024, %reg' is not a valid insn.

no regressions, except those from elimination of masks.

closes: #3

Signed-off-by: Charles E. Youse <charles@gnuless.org>