cc1: properly handle discarded regs in __asm()
lowering needs to account for the fact that some registers
in the defs clause of an I_ASM are discards and do not need
to be copied out. previously the compiler would crash because
it did not account for the REG_NONEs inserted by asm_stmt().

also made clarifying changes to regmap_regs(). these are
not functional changes, but make it clear that REG_NONE
is excluded from collection into the destination set.

Signed-off-by: Charles E. Youse <charles@gnuless.org>