cc1/func.c: decentralize decision to generate frame
rather than having logues() decide right at the end whether
the current function is going to need a frame, decide while
the function is being processed. we eliminate need_frame in
logues() and introduce a global func_needs_frame which is
then updated as necessary by frame_alloc() and arg0(). now,
by the time we get to logues(), the decision has been made...

... almost. logues() still makes a last-minute decision if
there are any non-volatile floating-point registers to save.
we'll rectify this separately in an upcoming commit.

groundwork to address bug #13 and issue #10. no regressions.

Signed-off-by: Charles E. Youse <charles@gnuless.org>