cc1: make I_MCH_RET* dependent on %rbp
to prevent the optimizer from trashing it.
also changed GP_K (and XMM_K along with it)
to constants, because the sizes of gp_colors
and xmm_colors will always be the same; also
need to add buffer room (+1) to GP_K since
RBP shows up in the interference graphs now,
just about everywhere.

fixes: #13

there is a lot of regression noise, but it's
just that. the changes in the interference
graphs cause the allocator to make slightly
different allocation and spill decisions.

Signed-off-by: Charles E. Youse <charles@gnuless.org>