cc1/dvn.c: preload frame with 0s
from the comments in dvn.c:

the data on the local frame (i.e., uninitialized variables) is undefined
upon function entry, so we are free to assume its state to be whatever we
want. we prepopulate the entry_block's reload state to appear as if the
entire frame had zeroes stored to it. well-formed programs will of course
not rely state of local variables before they're initialized, but this is
helpful when manipulating bitfields of locally-allocated structs. consider:

             f() { struct { int b : 2; int c : 30; } s; s.b = 1; }

ordinarily setting the field involves reading the entire word, manipulating
the bits in question [1:0] and then storing it back. obviously the state of
bits [31:2] is junk in this case, so assume they're zero, avoid the read,
and in this case everything collapses into a single store of a constant.

closes: #17

Signed-off-by: Charles E. Youse <charles@gnuless.org>