cc1: eliminate uses of __builtin_memcpy/set
these are going to be eliminated from the compiler.
it's far better to rely on the standard library to
do the Right Thing(tm) on ATOM, since it doesn't
have any fancy enhancements to its string insns.

this is a prelude to removing compiler support
for these builtins, and ultimately all builtins.
[__asm() is always there for do-it-yourselfers.]

Signed-off-by: Charles E. Youse <charles@gnuless.org>