kernel: introduce CURCPU and make sure BSP has APIC ID 0
it simplifies things if we can assume the BSP is APIC 0,
so let's do this unless and until we have a compelling
reason to do otherwise. check this on boot, panic if not.

never seen a real system in the wild for which this was
not true (though sometimes virtualization gets it wrong)

Signed-off-by: Charles E. Youse <charles@gnuless.org>