cc1/insn.c: clean up logic in insn_uses() and insn_defs()
previously some insns used `return' and others `break'.
this was an attempt to shortcut the fall-through logic
when it was a waste of time, but it can be confusing.

reposition the default logic to be the fall-through,
make sure that all cases fall out the bottoms of the
switches so that default logic has a chance to work.

no functional changes. no regressions.

Signed-off-by: Charles E. Youse <charles@gnuless.org>