kernel: ps/2 keyboard fixes
first, actually preserve the prior state before modifying it.
we were comparing the current state with junk, causing lots of
unnecessary LED updates to be scheduled, interfering with input.

second, clean up the KEY_* defines and the keymap[] formatting.

finally, add a few missing scan codes for the numeric keypad.

Signed-off-by: Charles E. Youse <charles@gnuless.org>