make: final touches. add to prototype filesystem.
make is complete, functions properly and, judging from
its long lifetime on minix with few changes, likely very
stable. so we're ready to move on from it for the moment.

unfortunately, the source code is fairly messy, and
the support for archive files will not work with our
custom archive format. issues will be opened for these.

Signed-off-by: Charles E. Youse <charles@gnuless.org>