compress: eliminate compile-time configuration
we're not portable, so int size, etc. are fixed,
and we have a large virtual-memory system. cut
the noise which tries to use memory availability
and data type sizes to decide preferred code size.

no changes to the binary. we've hardcoded the values
which would have otherwise been computed at compile.

Signed-off-by: Charles E. Youse <charles@gnuless.org>